WIP use PR ##1226 of vtk-js.
