+++ src/lib/loadImageData.js
@@ -1,8 +1,7 @@
 import cornerstone from 'cornerstone-core';
 import getSliceIndex from './data/getSliceIndex.js';
 import insertSlice from './data/insertSlice.js';
-
-const resolveStack = [];
+import getPatientWeightAndCorrectedDose from './data/getPatientWeightAndCorrectedDose.js';
 
 // TODO: If we attempt to load multiple imageDataObjects at once this will break.
 export default function loadImageDataProgressively(imageDataObject) {
@@ -14,6 +13,27 @@ export default function loadImageDataProgressively(imageDataObject) {
 
   const { imageIds, vtkImageData, metaDataMap, zAxis } = imageDataObject;
   const loadImagePromises = imageIds.map(cornerstone.loadAndCacheImage);
+  const imageId0 = imageIds[0];
+
+  const seriesModule = cornerstone.metaData.get(
+    'generalSeriesModule',
+    imageId0
+  );
+
+  if (!seriesModule) {
+    throw new Error('seriesModule metadata is required');
+  }
+
+  const modality = seriesModule.modality;
+
+  let modalitySpecificScalingParameters;
+
+  if (modality === 'PT') {
+    modalitySpecificScalingParameters = getPatientWeightAndCorrectedDose(
+      imageId0
+    );
+  }
+
   imageDataObject.isLoading = true;
 
   // This is straight up a hack: vtkjs cries when you feed it data with a range of zero.
@@ -21,16 +41,37 @@ export default function loadImageDataProgressively(imageDataObject) {
   const scalars = vtkImageData.getPointData().getScalars();
   const scalarData = scalars.getData();
 
-  console.log('SET SCALAR 1');
-
   scalarData[0] = 1;
+
+  const range = {
+    max: Number.NEGATIVE_INFINITY,
+    min: Number.POSITIVE_INFINITY,
+  };
 
   const insertPixelData = image => {
     return new Promise(resolve => {
       const { imagePositionPatient } = metaDataMap.get(image.imageId);
       const sliceIndex = getSliceIndex(zAxis, imagePositionPatient);
 
-      insertSlice(vtkImageData, sliceIndex, image);
+      const { max, min } = insertSlice(
+        vtkImageData,
+        sliceIndex,
+        image,
+        modality,
+        modalitySpecificScalingParameters
+      );
+
+      if (max > range.max) {
+        range.max = max;
+      }
+
+      if (min < range.min) {
+        range.min = min;
+      }
+
+      const dataArray = vtkImageData.getPointData().getScalars();
+
+      dataArray.setRange(range, 1);
 
       resolve();
     });
@@ -47,11 +88,6 @@ export default function loadImageDataProgressively(imageDataObject) {
   Promise.all(insertPixelDataPromises).then(() => {
     imageDataObject.isLoading = false;
     imageDataObject.loaded = true;
-
-    console.log('LOADED');
-    while (resolveStack.length) {
-      resolveStack.pop()();
-    }
 
     vtkImageData.modified();
   });
